0109
바닥 불 생성 구현
hp바 int 버그 수정
5층 ui 연동

+++ Assets/Scripts/UiManager.cs
@@ -88,7 +88,7 @@ public class UiManager : MonoBehaviour
             if (!bossBaseScript.isDie)
             {
                 bossHpText.text = bossBaseScript.currentHp + " / " + bossBaseScript.maxHp;
-                hpImage.fillAmount = bossBaseScript.currentHp / bossBaseScript.maxHp;
+                hpImage.fillAmount = bossBaseScript.currentHp / (float) bossBaseScript.maxHp;
             }
             else
                 hpGroup.SetActive(false);
